prod_requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
detoxify==0.2.2
discord.py==1.7.3
filelock==3.0.12
guppy3==3.1.1
huggingface-hub==0.0.12
idna==2.10
joblib==1.0.1
language-tool-python==2.5.4
multidict==5.1.0
nltk==3.6.2
numpy==1.21.0
packaging==20.9
protobuf==3.17.3
pyparsing==2.4.7
python-dotenv==0.18.0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
stanza==1.2.1
tokenizers==0.10.3
torch==1.9.0+cpu
tqdm==4.61.1
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
